% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_sbase_edaplant.R
\name{edaplant_sbase_server}
\alias{edaplant_sbase_server}
\title{Server EDA for SweetPotato Base}
\usage{
edaplant_sbase_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
server side for EDA for SweetPotato Base
}
\author{
Omar Benites
}

