% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_edaplant.R
\name{edaplant_server}
\alias{edaplant_server}
\title{Server EDA
Return the server component for exploratory graphics}
\usage{
edaplant_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
server side for EDA
}
\author{
Omar Benites
}

