% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_edaplant.R
\name{edaplant_ui}
\alias{edaplant_ui}
\title{UI for EDA
Returns a user interface for exploratory graphics}
\usage{
edaplant_ui(type = "tab", title = "Exploratory Data Analysis",
  name = "eda_plot_module")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
user interface for EDA
}
\author{
Omar Benites
}

