% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_ammi}
\alias{plot_ammi}
\title{AMMI and GGE}
\usage{
plot_ammi(model, biplot)
}
\arguments{
\item{model}{model}

\item{biplot}{1. genotype and environment effects. 2. principal components}
}
\description{
Only for MET data # Should run function ammi first
}
\author{
Raul Eyzaguirre
}

