% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_box}
\alias{plot_box}
\title{Boxplot and dotplot}
\usage{
plot_box(trait, by = NULL, dots = c("no", "yes"), fb)
}
\arguments{
\item{trait}{trait}

\item{by}{grouped by}

\item{dots}{use dot}

\item{fb}{field book data}
}
\description{
Plots two types of graphics: a boxplot and doplot.
}
\details{
Return a boxplot or dotplot chart
}
\author{
Omar Benites Raul Eyzaguirre
}
