% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_dens}
\alias{plot_dens}
\title{Density plot}
\usage{
plot_dens(trait, by = NULL, fb)
}
\arguments{
\item{trait}{trait}

\item{by}{grouped by}

\item{fb}{field book data}
}
\description{
Plots genotypes. Two options:  1. original values or 2, means over replications
}
\author{
Raul Eyzaguirre Omar Benites
}

