% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Scatterplot for more than two traits}
\usage{
plot_pairs(columns, fb)
}
\arguments{
\item{columns}{traits that we displayed in a dispersion matrix.}

\item{fb}{field book data}
}
\description{
Options: #1. Original values. #2. Means over replications.#3. Means over replications and environments.
}
\author{
Raul Eyzaguirre
}

