% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_scat}
\alias{plot_scat}
\title{Scatterplot for two traits}
\usage{
plot_scat(trait.1, trait.2, by = NULL, fb)
}
\arguments{
\item{trait.1}{first trait}

\item{trait.2}{second trait}

\item{by}{grouped by}

\item{fb}{field book data}
}
\description{
Options: 1. Original values, 2. Means over replications and  3. Means over replications and environments. 
With facets: One facet for each environment
}
\author{
Raul Eyzaguirre Omar Benites
}

