% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{shinyeda}
\alias{shinyeda}
\title{#' Adds the content of www to sF/
#' 
#' @importFrom shiny addResourcePath
#' 
#' @noRd
#' 
.onLoad <- function(...) {
  addResourcePath('sF', system.file('www', package='shinyFiles'))
}
Run a simple example app using the eda4plant functionality}
\usage{
shinyeda()
}
\description{
#' Adds the content of www to sF/
#' 
#' @importFrom shiny addResourcePath
#' 
#' @noRd
#' 
.onLoad <- function(...) {
  addResourcePath('sF', system.file('www', package='shinyFiles'))
}
Run a simple example app using the eda4plant functionality
}
