% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_hist}
\alias{plot_hist}
\title{Histogram for plant breeding traits}
\usage{
plot_hist(trait, bins, by = NULL, fb)
}
\arguments{
\item{trait}{trait}

\item{bins}{number of partitions or intervals.}

\item{by}{grouped by}

\item{fb}{field book data}
}
\description{
Plots two types of histogram: By trait and by trait x genotypes.
}
\author{
Raul Eyzaguirre Omar Benites
}

