% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plots_rhep.R
\name{plot_hist}
\alias{plot_hist}
\title{histogram}
\usage{
plot_hist(trait, bins, by = NULL, fb)

plot_hist(trait, bins, by = NULL, fb)
}
\arguments{
\item{trait}{trait}

\item{bins}{number of partitions or intervals.}

\item{by}{grouped by}

\item{fb}{datos}

\item{xcol}{x column}

\item{fb}{field book data}
}
\description{
plot the histogram

Plots genotypes. Two options:
}
\author{
Raul Eyzaguirre
}

