% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_scat}
\alias{plot_scat}
\title{Scatterplot for two traits}
\usage{
plot_scat(traitX, traitY, by = NULL, fb)
}
\arguments{
\item{traitX}{trait X}

\item{traitY}{trait Y}

\item{by}{grouped by}

\item{fb}{field book data}
}
\description{
Options: # 1. Original values, #2. Means over replications and  #3. Means over replications and environments. With facets: # One facet for each environment
}
\author{
Raul Eyzaguirre
}

