% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_rhep.R
\name{plot_scat}
\alias{plot_scat}
\title{Scatterplot for two traits}
\usage{
plot_scat(traitX, traitY, by = NULL, fb)
}
\arguments{
<<<<<<< HEAD
\item{traitX}{trait X}

\item{traitY}{trait Y}
=======
\item{trait.1}{first trait}

\item{trait.2}{second trait}
>>>>>>> 34e27d76d6132da2ca95bd214c14628965d606fa

\item{by}{grouped by}

\item{fb}{field book data}
}
\description{
Options: 1. Original values, 2. Means over replications and  3. Means over replications and environments. 
With facets: One facet for each environment
}
\author{
Raul Eyzaguirre Omar Benites
}
